Network and Site module basically OK

+++ ipview/handlers/admin.py
@@ -1,7 +1,7 @@
 import ipaddress
 from flask import Blueprint, render_template, url_for, redirect, flash, abort
-from ipview.forms import SiteForm, SupernetForm, AddSubnetForm, FilterForm
-from ipview.models import db, Site, Supernet, Subnet, IP, Event
+from ipview.forms import SiteForm, NetworkForm, AddSubnetForm, FilterForm
+from ipview.models import db, Site, Network, Subnet, IP, Event
 from flask_login import current_user
 
 
@@ -17,18 +17,18 @@ def log():
 
 @admin.route("/")
 def index():
-	return redirect(url_for("admin.supernet"))
+	return redirect(url_for("admin.network"))
 
 
-# site functions
+## site functions
 
 @admin.route("/site")
 def site():
     """display all sites
     """
-    sites = Site.query.order_by(Site.site_name).all()
+    sites = Site.query.order_by(Site.name).all()
     form = FilterForm()
-    form.filter_by.choices = [("site_name", "site name"), ("description", "description")]
+    form.filter_by.choices = [("name", "site name"), ("description", "description")]
     return render_template("admin/site.html", sites=sites, form=form)
 
 
@@ -41,7 +41,7 @@ def add_site():
     if form.validate_on_submit():
         form.populate_obj(site)
         if site.save():
-            site.log_event("Add new site: {}".format(site.site_name))
+            site.log_event("Add new site: {}".format(site.name))
             flash("success", "success")
         else:
             flash("failed to add site", "danger")
@@ -49,6 +49,14 @@ def add_site():
         return redirect(url_for("admin.site"))
     else:
         return render_template("admin/add_site.html", form=form)
+
+
+@admin.route("/site/<int:site_id>")
+def site_detail(site_id):
+    """display the subnets inside the site
+    """
+    site = Site.query.get_or_404(site_id)
+    return render_template("admin/site_detail.html", site=site)
 
 
 @admin.route("/site/<int:site_id>/edit", methods=['GET', 'POST'])
@@ -66,71 +74,84 @@ def edit_site(site_id):
 @admin.route("/site/<int:site_id>/delete")
 def delete_site(site_id):
     site = Site.query.get_or_404(site_id)
-    if site.delete():
-        site.log_event("Site is removed. site name: {}".format(site.site_name))
-        flash("success", "success")
+    if site.subnets == []:
+        if site.delete():
+            site.log_event("Removed site {}".format(site.name))
+            flash("site is removed", "success")
+        else:
+            flash("failed to remove this site", "danger")
     else:
-        flash("site remove failed", "danger")
+        flash("please remove the subnets under this site firstly", "danger")
 
     return redirect(url_for("admin.site"))
 
 
-# supernet functions
+## network functions
 
 @admin.route("/network")
-def supernet():
+def network():
     """display all networks(supernet CIDR)
     """
-    supernets = Supernet.query.all()
-    return render_template("admin/network.html", supernets=supernets)
+    networks = Network.query.all()
+    return render_template("admin/network.html", networks=networks)
 
 
 @admin.route("/network/add", methods=['GET', 'POST'])
-def add_supernet():
-    """add a new supernet
+def add_network():
+    """add a new network
     """
 
-    supernet = Supernet()
-    form = SupernetForm()
+    network = Network()
+    form = NetworkForm()
     if form.validate_on_submit():
-        form.populate_obj(supernet)
-        supernet.address_pack = int(ipaddress.ip_network(supernet.supernet_address).network_address)
-        if supernet.save():
-            supernet.log_event("Add a new supernet: {}".format(supernet.supernet_address))
-            flash("supernet added successfully", "success")
+        form.populate_obj(network)
+        network.address_pack = int(ipaddress.ip_network(network.address).network_address)
+        if network.save():
+            network.log_event("Add a new network: {}".format(network.address))
+            flash("network added successfully", "success")
         else:
-            flash("failed to add supernet", "danger")
-        return redirect(url_for("admin.supernet"))
+            flash("failed to add network", "danger")
+        return redirect(url_for("admin.network"))
     else:
-        return render_template("admin/add_supernet.html", form=form)
+        return render_template("admin/add_network.html", form=form)
 
 
-@admin.route("/network/<int:supernet_id>/edit")
-def edit_supernet(supernet_id):
-    """edit the supernet by supernet id
+@admin.route("/network/<int:network_id>/edit")
+def edit_network(network_id):
+    """edit the network by network id
     """
 
     pass
 
 
-@admin.route("/network/<int:supernet_id>/delete")
-def delete_supernet(supernet_id):
-    """delete the supernet by supernet id
+@admin.route("/network/<int:network_id>/delete")
+def delete_network(network_id):
+    """delete the network by supernet id
     """
 
-    pass
+    network = Network.query.get_or_404(network_id)
+    if network.subnets == []:
+        if network.delete():
+            network.log_event("Removed network {}".format(network.address))
+            flash("network is removed", "success")
+        else:
+            flash("failed to remove network", "danger")
+    else:
+        flash("please remove the sbunets under this network firstly", "danger")
+
+    return redirect(url_for("admin.network"))
 
 
-@admin.route("/network/<int:supernet_id>")
-def supernet_detail(supernet_id):
+@admin.route("/network/<int:network_id>")
+def network_detail(network_id):
     """display subnets inside the supernet
     """
 
-    supernet = Supernet.query.get_or_404(supernet_id)
-    return render_template("admin/network_detail.html", supernet=supernet)
+    network = Network.query.get_or_404(network_id)
+    return render_template("admin/network_detail.html", network=network)
 
 
-# subnet functions
+## subnet functions
 
 @admin.route("/subnet")
 def subnet():
@@ -140,37 +161,41 @@ def subnet():
     return render_template("admin/subnet.html", subnets=subnets)
 
 
-@admin.route("/network/<int:supernet_id>/add", methods=['GET', 'POST'])
-def add_subnet(supernet_id):
+@admin.route("/network/<int:network_id>/addsubnet", methods=['GET', 'POST'])
+def add_subnet_under(network_id):
     """add new subnet
     """
     subnet = Subnet()
-    supernet = Supernet.query.get_or_404(supernet_id)
+    network = Network.query.get_or_404(network_id)
     form = AddSubnetForm()
-    form.site_id.choices = [(site.id, site.site_name) for site in Site.query.order_by('site_name')]
+    form.site_id.choices = [(site.id, site.name) for site in Site.query.order_by('name')]
     if form.validate_on_submit():
         form.populate_obj(subnet)
-        subnet.supernet = supernet
-        subnet.address_pack = int(ipaddress.ip_network(subnet.subnet_address).network_address)
-        if subnet.save():
-            subnet_network = ipaddress.ip_network(subnet.subnet_address)    # ip_network obj for this subnet
-            for ip_addr in subnet_network.hosts():
+        subnet.network = network    # link to foreign key: network
+        subnet.address_pack = int(ipaddress.ip_network(subnet.address).network_address)
+        # after subnet created, write all its IP addresses into database IP table
+        if subnet.save():  
+            subnet_scope = ipaddress.ip_network(subnet.address)    # ip_network obj for this subnet
+            for ip_addr in subnet_scope.hosts():
                 ip = IP()
-                ip.ip_address = ip_addr.exploded
+                ip.address = ip_addr.compressed
                 ip.subnet = subnet
                 db.session.add(ip)
             try:
                 db.session.commit()
-                subnet.log_event("Add new subnent: {}".format(subnet.subnet_address))
             except:
                 db.session.rollback()
                 subnet.delete()
                 flash("failed to create IP addresses.", "danger")
+
+        if Subnet.query.filter_by(address=subnet.address):
+            subnet.log_event("Add new subnent: {}".format(subnet.address))
         else:
             flash("failed to add subnet", "danger")
-        return redirect(url_for("admin.network_detail", supernet_id=supernet.id))
+
+        return redirect(url_for("admin.network_detail", network_id=network.id))
     else:
-        return render_template("/admin/add_subnet.html", form=form, supernet=supernet)
+        return render_template("/admin/add_subnet.html", form=form, network=network)
 
 
 @admin.route("/subnet/<int:subnet_id>/edit")
@@ -193,7 +218,7 @@ def delete_subnet(subnet_id):
         db.session.rollback()
         flash("IP addresses removed failed", "danger")
     if subnet.delete():
-        subnet.log_event("Subnet {} is removed.".format(subnet.subnet_address))
+        subnet.log_event("Subnet {} is removed.".format(subnet.address))
         flash("success", "success")
     else:
         flash("subnet remove failed", "danger")
@@ -201,12 +226,16 @@ def delete_subnet(subnet_id):
     return redirect(url_for("admin.subnet"))
 
 
-@admin.route("/subnet/<int:subnet_id>")
-def subnet_detail(subnet_id):
+@admin.route("/<father>/<int:father_id>/subnet/<int:subnet_id>")
+def subnet_detail(father, father_id, subnet_id):
     """display all IP addresses in this subnet.
     """
+    if father == 'network':
+        father_obj = Network.query.get_or_404(father_id)
+    elif father == 'site':
+        father_obj = Site.query.get_or_404(father_id)
     subnet = Subnet.query.get_or_404(subnet_id)
-    return render_template("admin/subnet_detail.html", subnet=subnet)
+    return render_template("admin/subnet_detail.html", father=father, subnet=subnet)
